requirements.txt
websockets
python-telegram-bot==13.15